Organize better
