+++ crates/settings_content/src/settings_content.rs
@@ -3,16 +3,19 @@ mod fallible_options;
 mod git;
 pub mod merge_from;
 mod theme;
+mod ui;
 
 pub use editor::*;
 pub use fallible_options::*;
 pub use git::*;
 pub use merge_from::MergeFrom as MergeFromTrait;
 pub use theme::*;
+pub use ui::*;
 
 use serde::{Deserialize, Serialize};
-use settings_macros::{MergeFrom, with_fallible_options};
 use std::collections::HashMap;
+
+use settings_macros::{MergeFrom, with_fallible_options};
 
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub enum ParseStatus {
